task-management-system/backend/requirements/base.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1
